Weakening an overly-strong claim around `Sec-Nonsecure-Cookie`.
Reading this paragraph again, I find that it's broader than I intended. I do not think it's the case
that _no_ developer will react to our outreach, but I do think it's likely that a subset will be
surprised when we move from step 2 to step 3, breaking the link between HTTP and HTTPS
pages. I think there will be enough developers in this state to justify some temporary
bridge (whether we spell it `Sec-Nonsecure-Cookie` or in some other way), but I do not
think it will be as broad a group as the original text made it sound.